test-requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.
# Order matters to the pip dependency resolver, so sorting this file
# changes how packages are installed.  New dependencies should be
# added in alphabetical order, however, some dependencies may need to
# be installed in a specific order.
#
http://tarballs.openstack.org/horizon/horizon-master.tar.gz#egg=horizon
# Hacking should appear first in case something else depends on pep8
hacking<0.11,>=0.10.0
#
coverage>=3.6 # Apache-2.0
django-nose>=1.2 # BSD
flake8<2.6.0,>=2.5.4 # MIT
mock>=2.0 # BSD
mox3>=0.7.0 # Apache-2.0
networking-cisco>=3.1.0
nose # LGPL
nose-exclude # LGPL
nosehtmloutput>=0.0.3 # Apache-2.0
nosexcover # BSD
openstack.nose-plugin>=0.7 # Apache-2.0
oslosphinx!=3.4.0,>=2.5.0 # Apache-2.0
selenium>=2.50.1 # Apache-2.0
sphinx!=1.3b1,<1.3,>=1.2.1 # BSD
testtools>=1.4.0 # MIT
# This also needs xvfb library installed on your OS
xvfbwrapper>=0.1.3 #license: MIT
